+++ src/commands/libs/Collection.js
@@ -1,16 +1,15 @@
 import { authenticatedRequest } from '../../extra/utils';
 import { ServerResponse } from 'http';
 
-
 export class CollectionManager {
 	/**
-	 * 
-	 * @param {String} id 
-	 * 
+	 *
+	 * @param {String} id
+	 *
 	 * @returns {Collection} The collection
 	 */
 	static get(id) {
-		if (!id || typeof id !== 'string') {
+		if (!id || (typeof id !== 'string' && typeof id !== 'number')) {
 			throw new SyntaxError(`Expected id as String got ${typeof id}`);
 		}
 
@@ -18,13 +17,13 @@ export class CollectionManager {
 	}
 
 	/**
-	 * 
+	 *
 	 * @param {String} id Collection ID
-	 * 
+	 *
 	 * @returns {ServerResponse} The response object.
 	 */
 	static async delete(id) {
-		if (!id || typeof id !== 'string') {
+		if (!id || (typeof id !== 'string' && typeof id !== 'number')) {
 			throw new SyntaxError(`Expected id as String got ${typeof id}`);
 		}
 
@@ -32,21 +31,28 @@ export class CollectionManager {
 	}
 
 	/**
-	 * 
+	 *
 	 * @param {String} title Collection Title
 	 * @param {String} description Collection Description
 	 * @param {Boolean} isPrivate Is a private collection?
-	 * 
+	 *
 	 * @returns The created collection
 	 */
 	static async create(title, description, isPrivate = false) {
 		return await authenticatedRequest('/collections', {
 			method: 'POST',
-			body: JSON.stringify({
-				title,
-				description,
-				private: isPrivate
-			}, null, 2)
+			body: JSON.stringify(
+				{
+					title,
+					description,
+					private: isPrivate,
+				},
+				null,
+				2,
+			),
+			headers: {
+				'Content-Type': 'application/json',
+			},
 		});
 	}
 }
@@ -56,7 +62,6 @@ export class Collection {
 		this.id = id;
 	}
 
-
 	/**
 	 * @returns {Array} An array of photos.
 	 */
@@ -64,14 +69,12 @@ export class Collection {
 		return await authenticatedRequest(`/collections/${this.id}/photos`);
 	}
 
-
 	/**
 	 * @returns {Object} Collection.
 	 */
 	async info() {
 		return await authenticatedRequest(`/collections/${this.id}`);
 	}
-
 
 	/**
 	 * @returns An array of collections.
@@ -91,13 +94,14 @@ export class Collection {
 	 * @returns The added photo.
 	 */
 	async addPhoto(photo_id) {
-		if (!photo_id || typeof id !== 'string') {
+		if (!photo_id || (typeof photo_id !== 'string' && typeof id !== 'number')) {
 			throw new SyntaxError(`Expected id as String got ${typeof photo_id}`);
 		}
 
 		return await authenticatedRequest(`/collections/${this.id}/add`, {
 			method: 'POST',
-			body: JSON.stringify({ photo_id }, null, 2)
+			body: JSON.stringify({ photo_id }, null, 2),
+			json: true,
 		});
 	}
 
@@ -105,13 +109,13 @@ export class Collection {
 	 * @returns The removed photo.
 	 */
 	async removePhoto(photo_id) {
-		if (!photo_id || typeof photo_id !== 'string') {
+		if (!photo_id || (typeof photo_id !== 'string' && typeof id !== 'number')) {
 			throw new SyntaxError(`Expected id as String got ${typeof photo_id}`);
 		}
 
 		return await authenticatedRequest(`/collections/${this.id}/remove`, {
 			method: 'DELETE',
-			body: JSON.stringify({ photo_id }, null, 2)
+			body: JSON.stringify({ photo_id }, null, 2),
 		});
 	}
 }