+++ documentation/CHANGELOG.md
@@ -1,25 +1,40 @@
 # Changelog
+
+## 14 May 2019 (v3.4.0)
+
+- Fixed restore wallpaper
+- Fixed an Alias get error
+- Fixed Settings for last-wallpaper
+- Added Collection Management - You can now add and remove photos.
+- Better "help" menu for subcommands with the new `tableUtility`
+- Fixed some requests
+
 ## 21 November 2018 (v3.3.1)
+
 - Fixed prompt for like
 - Fixed `photo-of-the-day` delay check
 - Minor fixes
 
 ## 8 November 2018 (v3.2.1)
+
 - Removed `unsplash-js` library.
 - Removed `fetch` polyfill.
 - Custom `Unsplash` class.
 
 ### 3 November 2018 (v3.2.0)
+
 - Fixed like photo error
 - Changed **photo of the day** update.
 - Now the **photo of the day** update-delay can be configured in settings.
 - Implemented `jsbeautify`
 
 ### 1 November 2018 (v3.1.1)
+
 - Fixed [#35](/issues/35)
 - Improved user auth check.
 
 #### 22 October 2018 (v3.1.0)
+
 - Added default alias collection **Editorial** (317099).
 - Added login command (this will be implemented deeply in future versions).
 - Hidden all non user-useful settings from `settings get`.